requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
django-ratelimit==4.1.0
djangorestframework==3.16.0
Markdown==3.8
pytz==2025.2
sqlparse==0.5.3
